refactor(privacy): extract hardcoded JSX to MDX
- Move 364-line privacy page content into src/content/privacy.mdx
- Reduce src/app/privacy/page.tsx to 37-line thin shell using next-mdx-remote/rsc
- Add src/components/mdx/PrivacyComponents.tsx with FeaturesGrid, FeatureCard,
  ProcessorsGrid, ProcessorCard, ContactSection, and DataRightsForm components
- Typography and layout applied via custom MDX component mapping
- Legal content is now editable without a code deployment

Closes #1416 (privacy page — same pattern as terms MDX refactor)
